requirements.txt
pandas==1.3.3
tensorflow==2.7.0
keras==2.8.0
seaborn==0.11.2
numpy==1.21.3
matplotlib==3.4.3
opencv-python==4.5.3.56
flask==2.0.2
Pillow==8.4.0